[FEAT] VideoList UI 분리 및 ViewModel 적용
